refactor(batch): split test files by implementation file
coverage_test.go / contract_test.go → core_test.go, erc20_test.go,
stablecoin_test.go. Shared helpers and settings live in batch_test.go.
